requirements.txt
streamlit
google-generativeai
PyPDF2
fpdf==1.7.2
protobuf==4.25.3
